plenty of stuff

+++ db/clean_init_db.rb
@@ -1,4 +1,4 @@
-`rm db/schema.rb`
+`rm db/schema.rb` if File.exist?("db/schema.rb")
 `rails db:drop`
 `rails db:setup`
 `rails db:migrate`
